+++ lib/creator/index.js
@@ -3,8 +3,7 @@
 const yeoman = require("yeoman-environment");
 const Generator = require("yeoman-generator");
 const path = require("path");
-const defaultGenerator = require("./yeoman/webpack-generator");
-const WebpackAdapter = require("./yeoman/webpack-adapter");
+const defaultGenerator = require("../generators/init-generator");
 const runTransform = require("./transformations/index");
 
 /*
@@ -18,7 +17,7 @@ const runTransform = require("./transformations/index");
 */
 
 function creator(options) {
-	let env = yeoman.createEnv("webpack", null, new WebpackAdapter());
+	let env = yeoman.createEnv("webpack", null);
 	const generatorName = options
 		? replaceGeneratorName(path.basename(options[0]))
 		: "webpack-default-generator";
