+++ lib/generate-loader/index.js
@@ -1,5 +1,5 @@
 var yeoman = require("yeoman-environment");
-var LoaderGenerator = require("./loader-generator").LoaderGenerator;
+var LoaderGenerator = require("../generators/loader-generator").LoaderGenerator;
 
 /**
  * Runs a yeoman generator to create a new webpack loader project
