+++ lib/generators/init-generator.js
@@ -10,7 +10,7 @@ const Input = require("webpack-addons").Input;
 const Confirm = require("webpack-addons").Confirm;
 const RawList = require("webpack-addons").RawList;
 
-const getPackageManager = require("../../utils/package-manager")
+const getPackageManager = require("../utils/package-manager")
 	.getPackageManager;
 
 const entryQuestions = require("./utils/entry");
