+++ lib/generators/loader-generator.test.js
