+++ lib/generators/plugin-generator.js
@@ -1,6 +1,6 @@
 var path = require("path");
 var _ = require("lodash");
-var webpackGenerator = require("../utils/webpack-generator");
+var webpackGenerator = require("./webpack-generator");
 
 /**
  * A yeoman generator class for creating a webpack
