+++ lib/generators/utils/entry.js
