+++ lib/generators/utils/module.js
