+++ lib/generators/utils/plugins.js
