+++ lib/generators/utils/tooltip.js
