+++ lib/generators/utils/validate.js
