+++ lib/generators/webpack-generator.js
