+++ webpack.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-	entry: 'shit'
-}
+	entry: "shit"
+};
